refactor: change project folder structure
