+++ src/components/Loader/Loader.scss
