+++ src/components/MovieCard/MovieCard.module.scss
