+++ src/pages/MainPage/MainPage.module.scss
